[9] 캐싱 변경 작업 및 중복된 역 제거 및 가중치 변경 (#184)
* feature: 중복역 제거로직 추가

* feat: 캐싱 개선

+++ backend/goodday/src/main/java/seeuthere/goodday/location/domain/algorithm/TimeGrade.java
@@ -2,9 +2,9 @@ package seeuthere.goodday.location.domain.algorithm;
 
 public class TimeGrade {
 
-    private int minTime;
-    private int maxTime;
-    private int totalTime;
+    private double minTime;
+    private double maxTime;
+    private double totalTime;
 
     public TimeGrade() {
         minTime = Integer.MAX_VALUE;
@@ -12,17 +12,17 @@ public class TimeGrade {
         totalTime = 0;
     }
 
-    public void calculateTime(int time) {
+    public void calculateTime(double time) {
         minTime = Math.min(time, minTime);
         maxTime = Math.max(time, maxTime);
         totalTime += time;
     }
 
-    public int diffTime() {
+    public double diffTime() {
         return maxTime - minTime;
     }
 
-    public int getAvgTime(int size) {
+    public double getAvgTime(int size) {
         return totalTime / size;
     }
 }
